genereateQCreport_table

Generate QC Report Table for Seurat Object

Usage

generateQCreport_table(
  seu_obj = NULL,
  features = NULL,
  pdfFile = "QCReportTable.pdf",
  sample_id = NULL,
  path = NULL,
  platform = NULL
)

Arguments

seu_obj

A Seurat object, which must have properties like path, platform, and specific assay data. Must inherit from "Seurat".

features

Optional; A character vector of features to include in the analysis. Defaults to all features in seu_obj if NULL.

pdfFile

A character string specifying the path and name of the output PDF file. Defaults to "QCReportTable.pdf".

sample_id

Identifier for the sample being processed.

path

The file path from which to load the spatial transcriptomics data.

platform

The platform from which the data originates. Valid options are 'Xenium', 'CosMx', and 'Merscope'.

Value

A data frame with the QC metrics for the provided Seurat object, which is also printed to the console. Additionally, a PDF file is generated containing a plot and table of the metrics.

Details

Generate a QC report for a given Seurat object, allowing users to specify features and output to a PDF file. It provides various statistical analyses including entropy, sparsity, transcript counts per area, transcript counts per cell, and signal ratios. It also adjusts its behavior based on the specified platform within the Seurat object (Xenium, CosMx, or Merscope).

Calculate various metrics such as entropy and sparsity using the BioQC and coop packages. It adjusts its processing logic based on the 'platform' attribute of the Seurat object to handle data from different technologies like Xenium, CosMx, or Merscope. It stops with an error if the object is not a Seurat object. The results are plotted using ggplot2 and gridExtra for layout adjustments.